# executor, infoschema: add tiproxy into table `cluster_info`