<!--

Thank you for contributing to TiDB!

PR Title Format:
1. pkg [, pkg2, pkg3]: what's changed
2. *: what's changed

-->

### What problem does this PR solve?
<!--

Please create an issue first to describe the problem.

There MUST be one line starting with "Issue Number:  " and 
linking the relevant issues via the "close" or "ref".

For more info, check https://pingcap.github.io/tidb-dev-guide/contribute-to-tidb/contribute-code.html#referring-to-an-issue.

-->

Issue Number: ref #44827 

Problem Summary:
Add TiProxy to table `information_schema.cluster_info`.

### What is changed and how it works?
- Read TiProxy server info from `/topology/tiproxy/xxx/info`. The path for TiDB server info is `/tidb/server/info` but it's duplicated with `/topology/tidb/xxx/info`. I'll just use `/topology/tiproxy/xxx/info` for TiProxy.
- Add TiProxy to table `information_schema.cluster_info`.

### Check List

Tests <!-- At least one of them must be included. -->

- [x] Unit test
- [ ] Integration test
- [x] Manual test (add detailed scripts or steps below)
- [ ] No code

Before starting TiProxy:
```
mysql> select * from information_schema.cluster_info;
+------+-----------------+-----------------+---------+------------------------------------------+---------------------------+--------------+-----------+
| TYPE | INSTANCE        | STATUS_ADDRESS  | VERSION | GIT_HASH                                 | START_TIME                | UPTIME       | SERVER_ID |
+------+-----------------+-----------------+---------+------------------------------------------+---------------------------+--------------+-----------+
| tidb | 127.0.0.1:4000  | 127.0.0.1:10080 | 7.0.0   | 7376954cd868dbc44fc3015c9ef89c53749339a7 | 2023-06-20T15:43:14+08:00 | 2m5.673827s  |   2928446 |
| tidb | 127.0.0.1:4001  | 127.0.0.1:10081 | None    | None                                     | 2023-06-20T15:44:43+08:00 | 36.67383s    |   2921643 |
| pd   | 127.0.0.1:2379  | 127.0.0.1:2379  | 7.0.0   | da5a4e95ca0a739fc91e4dfd0f281e5413f5ff15 | 2023-06-20T15:40:59+08:00 | 4m20.673832s |         0 |
| tikv | 127.0.0.1:20160 | 127.0.0.1:20180 | 7.0.0   | 85e27ae5015694958a9bc123988775b2138ea58e | 2023-06-20T15:41:02+08:00 | 4m17.673835s |         0 |
+------+-----------------+-----------------+---------+------------------------------------------+---------------------------+--------------+-----------+
4 rows in set (0.00 sec)
```

After starting TiProxy:
```
mysql> select * from information_schema.cluster_info;
+---------+--------------------+--------------------+-------------------+------------------------------------------+---------------------------+------------------+-----------+
| TYPE    | INSTANCE           | STATUS_ADDRESS     | VERSION           | GIT_HASH                                 | START_TIME                | UPTIME           | SERVER_ID |
+---------+--------------------+--------------------+-------------------+------------------------------------------+---------------------------+------------------+-----------+
| tidb    | 127.0.0.1:4000     | 127.0.0.1:10080    | 7.0.0             | 7376954cd868dbc44fc3015c9ef89c53749339a7 | 2023-06-19T16:52:49+08:00 | 22h34m30.3175s   |   2221717 |
| tidb    | 127.0.0.1:4001     | 127.0.0.1:10081    | None              | None                                     | 2023-06-20T15:26:31+08:00 | 48.317505s       |   2786754 |
| pd      | 127.0.0.1:2379     | 127.0.0.1:2379     | 7.0.0             | da5a4e95ca0a739fc91e4dfd0f281e5413f5ff15 | 2023-06-19T16:52:30+08:00 | 22h34m49.317507s |         0 |
| tikv    | 127.0.0.1:20160    | 127.0.0.1:20180    | 7.0.0             | 85e27ae5015694958a9bc123988775b2138ea58e | 2023-06-19T16:52:34+08:00 | 22h34m45.317508s |         0 |
| tiproxy | 192.168.0.101:6000 | 192.168.0.101:3080 | refactor_infosync | 9ca256a0b53bfa45acba674025ffab73807a639e | 2023-06-20T15:25:09+08:00 | 2m10.317511s     |         0 |
+---------+--------------------+--------------------+-------------------+------------------------------------------+---------------------------+------------------+-----------+
5 rows in set (0.00 sec)
```

Side effects

- [ ] Performance regression: Consumes more CPU
- [ ] Performance regression: Consumes more Memory
- [ ] Breaking backward compatibility

Documentation

- [ ] Affects user behaviors
- [ ] Contains syntax changes
- [ ] Contains variable changes
- [ ] Contains experimental features
- [ ] Changes MySQL compatibility

### Release note

<!-- compatibility change, improvement, bugfix, and new feature need a release note -->

Please refer to [Release Notes Language Style Guide](https://pingcap.github.io/tidb-dev-guide/contribute-to-tidb/release-notes-style-guide.html) to write a quality release note.

Since TiProxy won't be released recently, it's invisible to users, so I won't add a release note here.

```release-note
None
```