@hawkingrei 2023-07-26
> @djshow832: The following test **failed**, say `/retest` to rerun all failed tests or `/retest-required` to rerun all mandatory failed tests:
> Test name 	Commit 	Details 	Required 	Rerun command
> tiprow_fast_test 	[a9cc4cf](https://github.com/pingcap/tidb/commit/a9cc4cf07cd9c59d076fa0decbfffc5d708c72f1) 	[link](https://tiprow.hawkingrei.com/view/gs/pingcapprow/pr-logs/pull/pingcap_tidb/44828/tiprow_fast_test/1683450409076658176) 	true 	`/test tiprow_fast_test`
> 
> [Full PR test history](https://tiprow.hawkingrei.com/pr-history?org=pingcap&repo=tidb&pr=44828). [Your PR dashboard](https://tiprow.hawkingrei.com/pr?query=is:pr+state:open+author:djshow832).

@djshow832 PTAL